small refactor; create strings file
